Create PasswordCheckerDomainUsers.ps1
This script checks the last password set date for domain user accounts by querying the domain controller. It reads a list of usernames from a specified text file and outputs the results to a CSV file.

In a domain environment, user accounts are managed on a domain controller, which allows for centralized control of user information and policies.

The /domain switch is necessary to retrieve details about users that are part of the Active Directory, ensuring that the script queries the correct user accounts.